pyproject.toml
[tool.poetry]
name = "memap"
version = "0.1.0"
description = "Gestão mapas mensais"
authors = ["mastervieira <javatang46@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.14, <3.15"
sqlalchemy = "^2.0.46"
pip = "^26.0"
pip-audit = "^2.10.0"
openpyxl = "^3.1.5"
reportlab = "^4.4.9"
pyside6 = "^6.10.1"
qasync = "^0.28.0"
pillow = "^12.1.1"
pypdf2 = "^3.0.1"
qtawesome = "^1.4.1"
pyqtgraph = "^0.14.0"


[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^6.0.0"
mypy = "^1.8.0"
pylint = "^3.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"
pytest-cov = "^6.0.0"
coverage = "^7.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ═══════════════════════════════════════════════════════════════════════════
# 🎨 BLACK - Formatação PEP8
# ═══════════════════════════════════════════════════════════════════════════
[tool.black]
line-length = 79
target-version = ['py314']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ═══════════════════════════════════════════════════════════════════════════
# 📦 ISORT - Organização de imports PEP8
# ═══════════════════════════════════════════════════════════════════════════
[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".venv", "venv", "__pycache__", "build", "dist"]

# ═══════════════════════════════════════════════════════════════════════════
# 🔍 MYPY - Type checking rigoroso
# ═══════════════════════════════════════════════════════════════════════════
[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
pretty = true
ignore_missing_imports = true

# Por módulo (exemplo)
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# ═══════════════════════════════════════════════════════════════════════════
# ✅ PYLINT - Análise de código PEP8
# ═══════════════════════════════════════════════════════════════════════════
[tool.pylint.main]
max-line-length = 79
good-names = ["i", "j", "k", "ex", "_", "id", "pk"]
disable = [
  "C0111",  # missing-docstring
  "C0103",  # invalid-name (para casos específicos)
  "R0903",  # too-few-public-methods
  "W0212",  # protected-access (para testes)
]

[tool.pylint.design]
max-attributes = 10
max-args = 7
max-locals = 15

[tool.pylint.similarities]
min-similarity-lines = 4

# ═══════════════════════════════════════════════════════════════════════════
# 🧪 PYTEST - Configuração de testes
# ═══════════════════════════════════════════════════════════════════════════
[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-v",
  "--strict-markers",
  "--tb=short",
  "--cov=src",
  "--cov-report=html",
  "--cov-report=term-missing:skip-covered",
]

# ═══════════════════════════════════════════════════════════════════════════
# 📊 COVERAGE - Cobertura de código
# ═══════════════════════════════════════════════════════════════════════════
[tool.coverage.run]
source = ["src"]
omit = [
  "*/tests/*",
  "*/test_*.py",
  "*/__pycache__/*",
  "*/venv/*",
  "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
]
